app/requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
rapidfuzz>=3.0.0
pymupdf>=1.24.0
pdf2image>=1.17.0
pytesseract>=0.3.13
pillow>=10.0.0
google-genai>=1.0.0
